alert

+++ src/components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, ScrollView, Text, FlatList } from 'react-native';
+import { View, ScrollView, Text, FlatList, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -29,7 +29,25 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite ?',
+                            "Are you sure want to delete the favorite dish " + item.name + " ?",
+                            [
+                                {
+                                    text: 'Cancle',
+                                    onPress: () => console.log(item.name + 'not deleted'),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id),
+                                    style: 'default'
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ];
             return (
